* small refactor common dropdown functions * rename function
