+++ src/client/src/components/Dropdown/index.tsx
@@ -3,6 +3,7 @@ import { ReactComponent as ArrowDown } from "../../assets/chevron.svg";
 
 import Select, { components } from "react-select";
 import dropdownstyles from "./dropdownstyles";
+import { ISelected } from "../../types/selected";
 
 import styles from "./styles.module.css";
 
@@ -30,6 +31,31 @@ const DropdownIndicator = (props: any) => {
     )
   );
 };
+
+export function convertOptionToDropdownItem(option: ISelected): IDropDownOptionType {
+  if (option.internalName && option.title) {
+    return {
+      value: option.internalName,
+      label: option.title,
+    };
+  }
+  return {
+    value: "",
+    label: "",
+  };
+}
+
+export function convertOptionsToDropdownItems(options: any[]): IDropDownOptionType[] {
+  const dropDownItems = [];
+  for (const option of options) {
+    if (option.unselectable) {
+      continue;
+    }
+    const dropdownItem = convertOptionToDropdownItem(option);
+    dropDownItems.push(dropdownItem);
+  }
+  return dropDownItems;
+}
 
 const Dropdown = ({
   options,
